Create gui
In the code given below, when the "clear" button of "x" is clicked, I want to clear only the file contents of "x" which is been chosen by the "choosexfile" that is been displayed by the "listbox" of the middleframe. And the same functionality must be done for "y". I tried using myList1.delete(0, 'end'), but it is clearing the contents entire file contents (both "x" & "y" file contents). But I want only the file contents of "x" to be cleared when the clear button of "x" is clicked & to clear the only the file contents of "y" to be cleared when the clear button of "y" is clicked. Can I use these two options: 1. Use ttk.Treeview instead of tk.Listbox, there is a tag parameter which allows to tag entry like tag='Soil' and delete by tag. 2. Create a variable list which holds the index of the inserted lines and delete by index. If yes may I know how? Iam a beginner so please be more clear.